I updated my API spec and my Database spec and implemented the changes
I split up the Assignment Endpoints to now have submission and will eventually implement feedback. I didn't want to have everything in one endpoint because it was getting more difficult to implement. I tried to do my best to test everything, but I have yet had the time to start a testing script. My next plan is to add the feedback endpoint and check over everything to make sure it not only lines up with my spec, but to make sure that my spec is what I actually want.