Updated the spec and added some requests into Feedback
I changed some of the functionality of the API (You can only get the latest submission, You can get all the feedback for a specific class, etc.). I also added put and patch requests into the feedback and plan to add removal of classes and assignments, but need to understand how to do it logistically without having holes in the database.