Testing Script for Semesters and Classes
I created two testing suites where I fully tested the Semester and Class API Calls (Any commented out code was involving the post method, so I removed it so as to prevent me continually adding more classes, but everything should still be in order as I checked). I also modified the code in all of the methods where I check for if a class or semester exists. This is to prevent any errors that could occur if there were holes in the database.