I fully tested the Assignment(s) endpoints
I also fixed small problems with my tracker_api.md and accordingly with my trackerAPI.py. I also made sure to move some test cases that were originally in the ClassMultiTestSuite into the SemesterMultiTestSuite which is where they belonged. Lastly, I fixed one of the error codes, so that it actually was correct for what the code ended up doing.