I added delete endpoints for Semesters, Classes, and Assignments and changed the way my post methods get their respective ids
I added delete methods in the Semester(Resource), but I needed to add RemoveClasses(Resource) and RemoveAssignments(Resource) to make the endpoints easier. I also changed the way I calculate the id in the post methods for all of my endpoints to make sure there are no repeat ids. This involves the way I error handle when determining whether or not an id is in the database to account for the holes.