Fully Tested the Semester, Class, and Assignments Resources in their respective test suites.
The ways in which I tested these before was extremely bad coding style and barely worked when I implemented the next suite. These tests work with the empty database, add and test everything, and then wipe the database clean at the end to guarantee the smooth operation of everything instead of working on a flimsy database of random elements.