Fixed Finding Assignments
I changed the way I checked if an assignment exists to match with the other ways as well as to prevent having issues when deleting assignments. If an assignment is deleted in the middle, and not exclusively the end, it will cause an error when trying to get it. I had previously added this functionality, but forgot to update it for my Feedbacks.